feat: add tests for locale-switcher.tsx
